requirements.txt
seaborn
matplotlib
scipy
numpy
unidecode
wikipedia
whoosh
nltk
scikit-learn
regex
fuzzywuzzy
six
py4j
python-Levenshtein
tornado
flask
requests
click
future
pyfunctional
fn
luigi